Rework layer construction interface
We'd like to be able to iterate over all available layers and instantiate
them. This is complicated by the fact that different layers need different
information; we handled it thus far by having the base Layer class have
explicit arguments (defaulting to None) for all potential data.

That's not a super flexible approach. We want to not only allow new
client-specific layers, but we also want to allow layers to apply to non-cfr
data (such as notice preambles). This patch allows this by allowing the base
class to receive arbitrary keyword args. So, when iterating over the layers,
we instantiate using all available contextual data -- layers that need some
of it define their required arguments at this point.

This will also explode if a layer is configured incorrectly. I'm imagining
trying to apply the CFR-specific "meta" layer to a notice preamble, which,
rightfully, makes no sense.